Refactor banner component to header; enhance responsive design and styling for home section
